requirements.txt
numpy
matplotlib
tifffile
scipy
pyfftw
torch
PyYAML
imagecodecs
scico
ray
ray[tune]
pandas
hyperopt


